[#12] Fetch updated timetables
